+++ CHANGELOG.md
@@ -2,16 +2,26 @@
 
 Here you will find all changes per version
 
+## v0.0.8
+
+**Enhancements**
+
+- See what version of database you have and see what the newest version is (No longer need to use -NewestVersion parameter, it does it automatically)
+- Show last time Database was checked for newest updates
+- Created a merge log that when there is a change it will show the changes that occurred
 
 ## v0.0.7
 
-**Enhancements
-	-** Ability to see your current database version and what is the newest version out
+**Enhancements**
+
+- Ability to see your current database version and what is the newest version out
 
 ## v0.0.6
 
-**Enhancements
-	-** Added database last updated field
+**Enhancements**
+
+- Added database last updated field
 
 **Fixed**
-	- JSON path for Keys now working in ./Assets/FidoKeys.json
+
+- JSON path for Keys now working in ./Assets/FidoKeys.json